Break out atomic constraint reader lookup for ioc
