+++ profile/build.gradle
@@ -4,6 +4,7 @@
 
 dependencies {
     compile project(':common')
+    compile 'com.google.inject:guice:4.1.0'
     compile 'org.jetbrains.kotlin:kotlin-stdlib:1.3.21'
     compile 'com.worldturner.medeia:medeia-validator-jackson:1.0.0'
     compile 'org.leadpony.justify:justify:0.14.0'
