+++ profile/src/test/java/com/scottlogic/deg/profile/reader/JsonProfileReaderTests.java
@@ -41,7 +41,9 @@ public class JsonProfileReaderTests {
 
     private Profile getResultingProfile() throws IOException, InvalidProfileException {
         if (this.profile == null) {
-            JsonProfileReader objectUnderTest = new JsonProfileReader();
+            JsonProfileReader objectUnderTest = new JsonProfileReader(
+                new MainConstraintReader(
+                    new BaseCatalogAtomicConstraintReaderLookup()));
             this.profile = objectUnderTest.read(this.json);
         }
